changes: Implemented changes to the last elr evaluation

+++ app/topic-response.go
@@ -157,17 +157,17 @@ func (tp *topicPartition) Encode() ([]byte, error){
 	}
 
 	//there might be an error in this section....can't find the elr from the source partition
-	if err := binary.Write(buff, binary.BigEndian, len(tp.lastKnowELR)); err != nil {
+	if err := binary.Write(buff, binary.BigEndian, int32(len(tp.lastKnowELR))); err != nil {
 		log.Printf("error while writing last known elr: %v\n", err)
 		return nil, err
 	}
 
-	if err := binary.Write(buff, binary.BigEndian, len(tp.lastKnowELR)); err != nil {
+	if err := binary.Write(buff, binary.BigEndian, int32(len(tp.lastKnowELR))); err != nil {
 		log.Printf("error while writing last know elr: %v\n", err)
 		return nil, err
 	}
 
-	if err := binary.Write(buff, binary.BigEndian, len(tp.offlineRepl)); err != nil {
+	if err := binary.Write(buff, binary.BigEndian, int32(len(tp.offlineRepl))); err != nil {
 		log.Printf("error while writing offline leader replica: %v\n", err)
 		return nil, err
 	}
